2.6.1 Land Use Classification/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
scikit-learn
jinja2
python-multipart
joblib
gunicorn
